+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+***Fixed:***
+
+- Workaround a limitation of the PTY dependency on Windows
+
 ## 0.14.0 - 2025-05-30
 
 ***Added:***